+++ sdk/nodejs/cosmosdb/index.ts
@@ -4,3 +4,4 @@
 // Export members:
 export * from "./account";
 export * from "./getAccount";
+export * from "./zMixins";